requirements.txt
amqp==1.4.9
apns2==0.3.0
asn1crypto==0.23.0
billiard==3.3.0.23
celery==3.1.25
certifi==2018.1.18
cffi==1.11.2
chardet==3.0.4
cryptography==2.1.4
Django==1.11.7
Django-api-base==0.10.2
django-celery==3.2.2
django-push-notifications==1.5.0
djangorestframework==3.7.3
googlemaps==2.5.1
gunicorn==19.7.1
h2==2.6.2
hpack==3.0.0
hyper==0.7.0
hyperframe==3.2.0
idna==2.6
kombu==3.0.37
psycopg2==2.7.3.2
pycparser==2.18
PyJWT==1.5.3
PySocks==1.6.8
pytz==2017.3
requests==2.18.4
rollbar==0.13.17
six==1.11.0
twilio==6.10.3
urllib3==1.22
vine==1.1.4
flower==0.9.2
gevent==1.3.7

